Make the PAL thread pool work for NTSC 2D/3D too.
See the comments added in decoder.h for a summary of the new design.
PalCombFilter becomes DecoderPool, with a generic Decoder interface;
the PAL FilterThread becomes PalDecoder, and NtscFilter is substantially
simplified into NtscDecoder.

getInputFrame now has to return the NTSC field phase IDs (which
PalDecoder ignores).

In main(), a generic -f option lets you select a decoder, with a
sensible default based on the input metadata. If you choose the ntsc3d
decoder (replacing -3), it forces the thread count to 1.