Compute the output height automatically, vertically centring the output.
This is exactly the same as we already do with the width: if we know the
active lines, the height is the next multiple of 8 above that, and we
extend the image by inserting blank lines -- as going outside the active
region vertically isn't a good idea -- switching between top and bottom
to keep the image centred.

In practice, the PAL size is already a multiple of 8, and NTSC needs to
insert 3 lines of padding.