Decode frames in batches.
Decoders now receive a QVector of fields to process, and return a
QVector of decoded frames. They can request that the input vector should
include an arbitrary amount of overlap at each end, which will be filled
with blank frames if it extends past the bounds of the file.

This has three advantages:

- 3D decoders, such as Comb in 3D mode or a future 3D FFT decoder,
  can now be parallelised. This is handled very simply for Comb by
  asking for one frame of history at the start of each batch, decoding
  it and throwing the result away; a 3D FFT would be given the vector of
  fields to read from directly.

- You can start a 3D decode in the middle of a file, and get a proper 3D
  result for the first frame.

- It reduces the cost of fetching the next frame to process by
  amortisation (which doesn't really make a big difference, as decoding
  a frame is relatively expensive anyway).