Remove refAmpl and refNorm.
The original PALcolour code needed refAmpl because it used scaled
integer arithmetic; we now use doubles throughout, so we can generate
the lookup table with amplitude 1 and not have to divide down after
multiplication by it.

Expanding out the multiplication of the composite signal by the two
carriers shows that the output of the filters is (U * refAmpl * 0.5, Vsw
* V * refAmpl * 0.5); refNorm was correcting for this after resynthesis.
I've moved the factor of 2 down since that's all that's left.

The U/V amplitude previously also had a scale factor of 2 -- but the
comment in the original code says "the '2' is arbitrary until I can work
out a physical reason for it". I've raised this to 2.56 to keep the
output the same for now, but this ought to depend on the difference
between white level and black level in the same way that the Y output
does.