Select replacement lines with the right phase for PAL.
Since there's a 90 degree phase shift per line, the line offset needs to
be different when moving from the first to second fields vs. the second
to first. This means findReplacementLine needs to know whether the field
it's correcting is the first field or the second field.

For NTSC, the 180 degree phase shift means it's the same either way.