Use different source lines for luma and chroma.
The chroma source still has to have the same chroma phase, but the luma
source can be an immediately adjacent line. This gives us more choices
of luma replacements and means we'll usually find one that looks better.

The filtering is a bit inefficient at the moment because it filters two
complete lines for each replacement; it might be worth caching the
filter result to improve performance on very dropout-y sources.