Use QVector<quint16> rather than QByteArray for TBC fields.
This avoids needing to reinterpret_cast to quint16 in all the places
that work with field data; the only unsafe casts are now when the data
is being read from or written to a file. It also avoids a few
unnecessary copies of the data (e.g. several places where a non-const
pointer was taken).

It also fixes a few minor problems:

- A couple of functions explicitly treated the TBC data as
  little-endian; it's now native-endian throughout.

- In ld-dropout-correct, statistics could be reported for the wrong
  frame, because they were read from the new OutputFrame rather than the
  one being written out.

- In ld-process-vbi, WhiteFlag was stepping by bytes rather than words
  through the samples (but since the samples should be either black or
  white, this shouldn't really make much difference).