Add ld-chroma-decoder --simple-pal for a 1D UV filter.
This gives better vertical resolution, at the cost of not dealing with
chroma phase errors. Thanks to Richard Russell for providing the specs
for the filter from the BBC's original Transform PAL implementation.

This only works with transform2d/3d at the moment, but in the future
it'd be possible to add a luma notch filter, giving a pal1d mode.