Replace JsonWax with an event-based parser.
Previously LdDecodeMetaData stored the metadata internally using
JsonWax's tree representation, which is very inefficient -- it doesn't
know the types of any of the data -- and made both parsing and doing
operations on it very expensive.

The TBC library now includes JsonReader and JsonWriter, which provide a
very simple event-based JSON parser. (It doesn't handle \u escapes
because that means doing gymnastics with UTF-8 encoding, but given how
little ld-decode uses strings this is probably not a problem for now.)

LdDecodeMetaData now reads data directly into the structures that the
tools use. In particular, there is now a vector of Field objects which
are returned by const reference, rather than having it laboriously
construct a Field from the JSON data every time it's requested.

The only downside of this is that LdDecodeMetaData can no longer
preserve members in JSON objects that it doesn't know about.

For a typical CLV disc, testmetadata now takes 1.1s rather than 7.1s.