Convert LineParameters defaults into structs.
Change LineParameters::process into LineParameters::applyTo, which takes
a reference to a VideoParameters struct to modify. This means that it
knows what video standard is in use and doesn't have to guess based on
frameHeight.

We can then select a set of defaults from a struct for each standard,
rather than using lots of conditional expressions. (This will also
simplify adding PAL-M support in the future, for vhs-decode.)