Support multiple video systems, and add PAL-M support (#728)
* Improve support for vhs-decode _chroma.tbc files in ld-analyse.

When loading a .tbc file, ld-analyse will now check whether the
corresponding _chroma.tbc file exists. If so, it'll load both of them,
and give you the option of whether to view luma, chroma, or the sum of
the two. (It doesn't decode them separately; that would be doable but
a bit more complicated.)

If you open a _chroma.tbc directly, it'll try to open both files if they
exist, defaulting to the chroma view; otherwise it'll look for the luma
JSON as before.

* Allow more than two video systems, and add support for PAL-M.

In the JSON, isSourcePal is replaced by a "system" string. In the
library, this is converted to a VideoSystem enum; if "system" is not
set, then the system is guessed from isSourcePal and fieldHeight, so the
tools will still work with old files.

PAL-M uses PAL colour on a 525-line line standard with fSC being close
to but not the same as NTSC -- so in some senses it works like standard
PAL, and in some senses like standard NTSC. Adjust all the
system-dependent parts of the tools as needed, including filters. This
was tested with a VHS tape decoded by vhs-decode.

I've assumed that closed captions aren't supported on PAL-M, although
there's no particular reason why they couldn't be; if this is needed in
the future then we should probably move the CC fields out of the NTSC
structure.