Add --field-offset option to ld-chroma-encoder.
Previously, the first output frame would always be field 0 in the NTSC
or PAL colour sequence. However, if you're trying to match a video
that's previously been encoded (e.g. to compare with a recording of a
different encoder), it's more likely than not that you'll need to start
at a different point in the sequence.