Write NTSC metadata to JSON more compactly.
Don't write ccData0/1 if they're not valid, the white flag if it's not
present, or fieldFlag if the FM code isn't valid. All of these have
sensible default values when they're read back in.

Also, despite my comment saying to keep the members in alphabetical
order, they weren't -- now they are.