Move CC data into its own JSON object.
Previously it was in the "ntsc" object (LdDecodeMetaData::Ntsc) -- but
CC is used on some PAL material as well. This moves it to a new "cc"
object, with the "ntsc" parser still recognising the old location for
compatibility with existing JSON files.

Remove some assumptions that CC is only for NTSC.